massive backend updates
the txt2img and img2img dropdowns are now separated. each has its own config file and its own config presets.
adjusted to sampler index not using checkboxes, but a dropdown now.
adjusted for the removal of firstpass width/height, and added Upscale by.
components on in the UI are tracked with elem_id instead of labels.
when creating a new preset, need to restart the UI because of stupid Gradio desync bug.
moved our dropdown UI section to below the gallery, instead of below the gallery's buttons.
added error handling incase the A1111 UI changes the again to hopefully not brick people's UI.
Create button renamed to Save & Restart.
#16