Reworked tracked IDs file generation
I moved the contents of both txt and img2img tracked id files to their own separate template files. This was done for two reasons, to help make it easier to add new IDs for users and to ensure the next part works.

Most of the new code is aimed at being able to properly append old existing files with new information in the template files. Now if you update the template files, user's files will be appended with new lines on their next reload. It will not overwrite existing data, it will only append new lines to the end of old files.

Moving the files was neccessary to accomplish this as the script needs to be able to read each file line by line and as far as I know that can't be copied with the old way these files were generated.